+++ templates-remotion/src/index.ts
@@ -1,4 +1,7 @@
 import { registerRoot } from "remotion";
 import { Root } from "./Root";
+import { registerCustomFonts } from "./fonts";
+
+registerCustomFonts();
 
 registerRoot(Root);